+++ tests/mongoc/test-mongoc-gridfs-file-page.c
@@ -112,7 +112,7 @@ test_read (void)
    ASSERT (page->offset == 3);
 
    r = _mongoc_gridfs_file_page_read (page, buf, 50);
-   ASSERT (r == len - 3);
+   ASSERT (r == (int32_t)len - 3);
    ASSERT (memcmp (fox + 3, buf, len - 3) == 0);
 
    _mongoc_gridfs_file_page_destroy (page);
